+++ app/components/Placeholders/IssueViewPlaceholder/styled/index.jsx
@@ -1,4 +1,4 @@
-import styled from 'styled-components2';
+import styled from 'styled-components';
 
 export const IssueViewPlaceholderContainer = styled.div`
   display: flex;