+++ app/containers/AuthForm/AuthDebugger/index.jsx
@@ -1,0 +1,4 @@
+// @flow
+import AuthDebugger from './AuthDebugger';
+
+export default AuthDebugger;