+++ app/containers/AuthForm/AuthForm.jsx
@@ -42,6 +42,9 @@ import {
 
 import EmailStep from './EmailStep';
 import TeamStep from './TeamStep';
+import AccountsStep from './AccountsStep';
+
+import AuthDebugger from './AuthDebugger';
 
 import {
   Hint,
@@ -49,40 +52,65 @@ import {
   Container,
   Logo,
   LoginInfo,
+  AuthDebugContainer,
 } from './styled';
 
 type Props = {
+  showAuthDebugConsole: boolean,
   loginRequestInProcess: boolean,
   loginError: string,
   isPaidUser: boolean,
+  accounts: Array<{| host: string, username: string |}>,
   host: string | null,
   step: number,
   dispatch: Dispatch,
 } & FormProps
 
 const AuthForm: StatelessFunctionalComponent<Props> = ({
+  showAuthDebugConsole,
   loginRequestInProcess,
   loginError,
   isPaidUser,
+  accounts,
   host,
   step,
   handleSubmit,
   dispatch,
 }: Props): Node =>
   <Container>
-    <Logo src={logoShadowed} alt="Chronos" />
+    <AuthDebugger show={showAuthDebugConsole} />
+    <Logo
+      src={logoShadowed}
+      alt="Chronos"
+    />
     <Flex column alignCenter>
       <LoginInfo>
         Log in to your account
       </LoginInfo>
       <ContentOuter>
+        {accounts.length > 0 &&
+          <AccountsStep
+            isActiveStep={step === 0}
+            dispatch={dispatch}
+            accounts={accounts}
+            onContinue={handleSubmit((data) => {
+              dispatch(authActions.loginRequest(data));
+            })}
+            loginError={loginError}
+            onBack={() => {
+              dispatch(uiActions.setUiState('authFormStep', 1));
+            }}
+          />
+        }
         <TeamStep
           isActiveStep={step === 1}
+          accounts={accounts}
           onContinue={() => {
             dispatch(
               uiActions.setUiState('authFormStep', 2),
             );
           }}
+          dispatch={dispatch}
           loginError={loginError}
         />
         <EmailStep
@@ -106,14 +134,25 @@ const AuthForm: StatelessFunctionalComponent<Props> = ({
         />
       </ContentOuter>
     </Flex>
-    <Hint>Can not log in?</Hint>
+    <Hint
+      onClick={() => {
+        dispatch(uiActions.setUiState(
+          'showAuthDebugConsole',
+          true,
+        ));
+      }}
+    >
+      Can not log in?
+    </Hint>
   </Container>;
 
 const selector = formValueSelector('auth');
 function mapStateToProps(state) {
   return {
+    showAuthDebugConsole: getUiState('showAuthDebugConsole')(state),
     host: selector(state, 'host'),
     step: getUiState('authFormStep')(state),
+    accounts: getUiState('accounts')(state),
     loginError: getUiState('loginError')(state),
     loginRequestInProcess: getUiState('loginRequestInProcess')(state),
     // Temporary block OAuth