+++ app/main.dev.js
@@ -1,5 +1,6 @@
 /* eslint-disable no-console, global-require */
 import path from 'path';
+import fs from 'fs';
 import keytar from 'keytar';
 import storage from 'electron-json-storage';
 import {
@@ -8,16 +9,22 @@ import {
   Menu,
   MenuItem,
   ipcMain,
+  clipboard,
   BrowserWindow,
   screen,
   session,
+  dialog,
 } from 'electron';
 import notifier from 'node-notifier';
 import MenuBuilder from './menu';
+import config from './config';
+
+import pjson from '../package.json';
 
 const appDir = app.getPath('userData');
 
 let mainWindow;
+let issueWindow;
 let tray;
 let menu;
 let authWindow;
@@ -219,6 +226,9 @@ function createWindow(callback) {
           }
         } else if (process.platform === 'darwin') {
           checkRunning(e);
+          if (shouldQuit) {
+            issueWindow.destroy();
+          }
         }
       }
     });
@@ -392,46 +402,76 @@ ipcMain.on('issue-refetch', (event, issueId) => {
   mainWindow.webContents.send('reFetchIssue', issueId);
 });
 
+ipcMain.on('save-login-debug', (event, messages) => {
+  const log = messages.map(message => (message.string
+    ? `${message.string}`
+    : `${JSON.stringify(message.json, null, 2)}`
+  )).join('\n');
+  dialog.showSaveDialog(mainWindow, {
+    defaultPath: `chronos-${pjson.version}-auth-debug.log`,
+  },
+  (filename) => {
+    if (filename) {
+      fs.writeFileSync(filename, log);
+    }
+  });
+});
+
+ipcMain.on('copy-login-debug', (event, messages) => {
+  const log = messages.map(message => (message.string
+    ? `${message.string}`
+    : `${JSON.stringify(message.json, null, 2)}`
+  )).join('\n');
+  clipboard.writeText(log);
+});
+
 ipcMain.on('load-issue-window', (event, url) => {
-  const createIssueWindow = new BrowserWindow({
-    backgroundColor: 'white',
-    parent: mainWindow,
-    show: false,
-    modal: true,
-    useContentSize: true,
-    center: true,
-    title: 'Chronos',
-    webPreferences: {
-      devTools: true,
-    },
-  });
-  createIssueWindow.loadURL(`file://${__dirname}/issueForm.html`);
-  createIssueWindow.webContents.on('did-finish-load', () => {
-    createIssueWindow.webContents.send('url', url);
-  });
-  if (process.env.NODE_ENV === 'development' || process.env.DEBUG_PROD === true) {
-    createIssueWindow.openDevTools();
+  if (!issueWindow) {
+    issueWindow = new BrowserWindow({
+      backgroundColor: 'white',
+      parent: mainWindow,
+      show: false,
+      modal: true,
+      useContentSize: true,
+      center: true,
+      title: 'Chronos',
+      webPreferences: {
+        devTools: true,
+      },
+    });
+    issueWindow.loadURL(`file://${__dirname}/issueForm.html`);
+    issueWindow.webContents.on('did-finish-load', () => {
+      issueWindow.webContents.send('url', url);
+    });
+    if (
+      config.issueWindowDevTools ||
+      process.env.DEBUG_PROD === true
+    ) {
+      issueWindow.openDevTools();
+    }
+    issueWindow.on('close', (cEv) => {
+      cEv.preventDefault();
+      issueWindow.hide();
+    });
+    ipcMain.on('page-fully-loaded', () => {
+      if (issueWindow) {
+        issueWindow.webContents.send('page-fully-loaded');
+      }
+    });
+    ipcMain.on('close-issue-window', () => {
+      if (issueWindow) {
+        issueWindow.hide();
+      }
+    });
+    ipcMain.on('show-issue-window', (ev, opts) => {
+      if (issueWindow) {
+        issueWindow.webContents.send('showForm', opts);
+        issueWindow.show();
+      }
+    });
+  } else {
+    issueWindow.webContents.send('url', url);
   }
-  createIssueWindow.on('close', (cEv) => {
-    cEv.preventDefault();
-    createIssueWindow.hide();
-  });
-  ipcMain.on('page-fully-loaded', () => {
-    if (createIssueWindow) {
-      createIssueWindow.webContents.send('page-fully-loaded');
-    }
-  });
-  ipcMain.on('close-issue-window', () => {
-    if (createIssueWindow) {
-      createIssueWindow.hide();
-    }
-  });
-  ipcMain.on('show-issue-window', (ev, opts) => {
-    if (createIssueWindow) {
-      createIssueWindow.webContents.send('showForm', opts);
-      createIssueWindow.show();
-    }
-  });
 });
 
 ipcMain.on('oauth-response', (event, text) => {