+++ app/sagas/auth.js
@@ -13,6 +13,7 @@ import {
 import Raven from 'raven-js';
 
 import * as Api from 'api';
+import * as R from 'ramda';
 
 import {
   actionTypes,
@@ -22,6 +23,7 @@ import {
 
 import {
   setToStorage,
+  getFromStorage,
   removeFromStorage,
 } from './storage';
 import {
@@ -29,6 +31,7 @@ import {
 } from './initializeApp';
 import {
   throwError,
+  notify,
 } from './ui';
 import createIpcChannel from './ipc';
 
@@ -99,12 +102,53 @@ export function* chronosBackendAuth({
   }
 }
 
+function storeInKeytar(payload, host) {
+  ipcRenderer.sendSync(
+    'store-credentials',
+    {
+      ...payload,
+      host: host.hostname,
+    },
+  );
+}
+
+function* saveAccount(payload: { host: string, username: string }): Generator<*, void, *> {
+  const { host, username } = payload;
+  let accounts = yield call(getFromStorage, 'accounts');
+  if (!accounts) accounts = [];
+  if (!R.find(R.whereEq({ host, username }), accounts)) {
+    accounts.push(payload);
+    yield call(setToStorage, 'accounts', accounts);
+  }
+}
+
+function* deleteAccount(payload: { host: string, username: string }): Generator<*, void, *> {
+  const { host, username } = payload;
+  let accounts = yield call(getFromStorage, 'accounts');
+  if (!accounts) accounts = [];
+  const index = R.findIndex(R.whereEq({ host, username }), accounts);
+  if (index !== -1) {
+    accounts = R.without([{ host, username }], accounts);
+    yield call(setToStorage, 'accounts', accounts);
+  }
+}
+
 export function* basicAuthLoginForm(): Generator<*, void, *> {
   while (true) {
     try {
       const { payload } = yield take(actionTypes.LOGIN_REQUEST);
       const host = yield call(transformValidHost, payload.host);
       const protocol = host.protocol.slice(0, -1);
+
+      yield put(authActions.addAuthDebugMessage([
+        { string: 'Login request...' },
+        { string: `host: ${host.hostname}` },
+        { string: `protocol: ${protocol}` },
+        { string: `port: ${host.port}` },
+        { string: `path_prefix: ${host.pathname}` },
+        { string: `username: ${payload.username}` },
+        { string: 'password: ***' },
+      ]));
 
       yield put(uiActions.setUiState('loginRequestInProcess', true));
       yield put(uiActions.setUiState('loginError', null));
@@ -119,7 +163,14 @@ export function* basicAuthLoginForm(): Generator<*, void, *> {
         },
       );
       // Test request for check auth
-      const userData = yield call(Api.jiraProfile);
+      const {
+        debug,
+        result,
+      } = yield call(Api.jiraProfile, true);
+      yield put(authActions.addAuthDebugMessage([
+        { json: debug },
+      ]));
+      const userData = result;
       if (!userData.self || !userData.active) {
         Raven.captureMessage('Strange auth response!', {
           level: 'error',
@@ -141,7 +192,14 @@ export function* basicAuthLoginForm(): Generator<*, void, *> {
       */
       yield call(
         setToStorage,
-        'jira_credentials',
+        'last_used_account',
+        {
+          username: payload.username,
+          host: payload.host,
+        },
+      );
+      yield call(
+        saveAccount,
         {
           username: payload.username,
           host: payload.host,
@@ -154,27 +212,26 @@ export function* basicAuthLoginForm(): Generator<*, void, *> {
           protocol,
         },
       );
-      yield call(
-        (): void => {
-          ipcRenderer.sendSync(
-            'store-credentials',
-            {
-              ...payload,
-              host: host.hostname,
-            },
-          );
-        },
-      );
+      yield call(storeInKeytar, payload, host);
       yield put(uiActions.setUiState('loginRequestInProcess', false));
       trackMixpanel('Jira login');
       incrementMixpanel('Jira login', 1);
     } catch (err) {
+      if (err.debug) {
+        err.debug.options.auth.password = '***';
+        err.debug.request.headers.authorization = '***';
+        yield put(authActions.addAuthDebugMessage([
+          {
+            json: err.debug,
+          },
+        ]));
+      }
       yield put(uiActions.setUiState('loginRequestInProcess', false));
       yield put(uiActions.setUiState(
         'loginError',
         'Can not authenticate user. Please try again',
       ));
-      yield call(throwError, err);
+      yield call(throwError, err.result ? err.result : err);
     }
   }
 }
@@ -259,7 +316,7 @@ export function* oAuthLoginForm(): Generator<*, *, *> {
 
 export function* logoutFlow(): Generator<*, *, *> {
   while (true) {
-    yield take(actionTypes.LOGOUT_REQUEST);
+    const { payload: { dontForget } } = yield take(actionTypes.LOGOUT_REQUEST);
     try {
       const { getGlobal } = remote;
       const { running, uploading } = getGlobal('sharedObj');
@@ -272,13 +329,18 @@ export function* logoutFlow(): Generator<*, *, *> {
         // eslint-disable-next-line no-alert
         window.alert('Currently app in process of saving worklog, wait few seconds please');
       }
-      if (!running && !uploading) {
+      if (!running && !uploading && !dontForget) {
         yield call(removeFromStorage, 'desktop_tracker_jwt');
-        yield call(removeFromStorage, 'jira_credentials');
+        const lastUsedAccount = yield call(getFromStorage, 'last_used_account');
+        yield call(deleteAccount, lastUsedAccount);
+        yield call(removeFromStorage, 'last_used_account');
       }
       yield put({
         type: actionTypes.__CLEAR_ALL_REDUCERS__,
       });
+      let accounts = yield call(getFromStorage, 'accounts');
+      if (!accounts) accounts = [];
+      yield put(uiActions.setUiState('accounts', accounts));
       trackMixpanel('Logout');
       incrementMixpanel('Logout', 1);
     } catch (err) {
@@ -310,6 +372,72 @@ function getOauthChannelListener(channel, type) {
       }
     }
   };
+}
+
+export function* switchAccountFlow(): Generator<*, *, *> {
+  while (true) {
+    const { payload } = yield take(actionTypes.SWITCH_ACCOUNT);
+    try {
+      const { getGlobal } = remote;
+      const { running, uploading } = getGlobal('sharedObj');
+
+      if (running) {
+        // eslint-disable-next-line no-alert
+        window.alert('Tracking in progress, save worklog before logout!');
+      }
+      if (uploading) {
+        // eslint-disable-next-line no-alert
+        window.alert('Currently app in process of saving worklog, wait few seconds please');
+      }
+      if (!running && !uploading) {
+        const host = yield call(transformValidHost, payload.host);
+        const {
+          credentials,
+          error,
+        } = ipcRenderer.sendSync(
+          'get-credentials',
+          {
+            username: payload.username,
+            host: host.hostname,
+          },
+        );
+        if (error) {
+          Raven.captureMessage('keytar error!', {
+            level: 'error',
+            extra: {
+              error: error.err,
+            },
+          });
+          yield call(
+            throwError,
+            error.err,
+          );
+          if (error.platform === 'linux') {
+            yield fork(
+              notify,
+              {
+                type: 'libSecretError',
+                autoDelete: false,
+              },
+            );
+          }
+        } else {
+          yield put({
+            type: actionTypes.__CLEAR_ALL_REDUCERS__,
+          });
+          yield put(uiActions.setUiState('initializeInProcess', true));
+          let accounts = yield call(getFromStorage, 'accounts');
+          if (!accounts) accounts = [];
+          yield put(uiActions.setUiState('accounts', accounts));
+          yield put(authActions.loginRequest({ ...payload, password: credentials.password }));
+        }
+      }
+      trackMixpanel('SwitchAccounts');
+      incrementMixpanel('SwitchAccounts', 1);
+    } catch (err) {
+      yield call(throwError, err);
+    }
+  }
 }
 
 export function* createIpcAuthListeners(): Generator<*, *, *> {