+++ app/styles/index.jsx
@@ -1,6 +1,5 @@
 // @flow
-import styled from 'styled-components2';
-
+import styled from 'styled-components';
 
 export const AppWrapper = styled.div`
   height: 100%;