+++ app/types/ui.js
@@ -108,6 +108,7 @@ export type UiState = {|
     confirmDeleteWorklog: boolean,
     settings: boolean,
     worklog: boolean,
+    accounts: boolean,
   |},
 
   flags: Array<any>,